Improve check from_lot timestamp order check
GainLossSet._sort_entries() performs a check that entry.from_lot.timestamp is > than parent.from_lot_timestamp. However if parent is a EARN transaction it doesn't have a from_lot. So the code doesn't perform this check. However there could be an ancestor with a from_lot further up the parent chain: the code should check against that ancestor (if it exists).